requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.4
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0

# Technical Analysis
TA-Lib==0.4.28
pandas-ta==0.3.14b0

# Backtesting
backtrader==1.9.78.123
vectorbt==0.25.2

# Data fetching
yfinance==0.2.32
nsepy==1.0.7

# Upstox API
upstox-python-sdk==2.0.0

# Angel One API (alternative)
smartapi-python==1.3.0

# Utilities
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
loguru==0.7.2

# Database (optional for trade logging)
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
